refactor: 숙소 관련 API URI 변경
